コンポーザブルからユーティリティ関数に変更
